feat: move getTranslatedSortFilterLabel function to sorting.ts and update its implementation
