+++ src/helpers/sorting.ts
@@ -3,11 +3,21 @@
  *
  * SPDX-License-Identifier: AGPL-3.0-only
  */
-import { updateSettings } from '@zextras/carbonio-shell-ui';
+import { updateSettings, useUserSettings } from '@zextras/carbonio-shell-ui';
 import { isTrash, JSNS } from '@zextras/carbonio-ui-commons';
 import { AccountSettingsPrefs, soapFetchV2 } from '@zextras/carbonio-ui-soap-lib';
 
-import type { FolderSortOrder, SortDirection } from '../types';
+import type {
+	FilterOption,
+	FolderSortOrder,
+	SortAndFilterState,
+	SortDirection,
+	SortOption
+} from '../types';
+import { FILTER_OPTIONS, SORTING_DIRECTION, SORTING_OPTIONS } from '../constants';
+import React, { useMemo } from 'react';
+import { useTranslation } from 'react-i18next';
+import { TFunction } from 'i18next';
 
 /**
  * Returns sortType, sortDirection and sortOrder for the given folder
@@ -133,3 +143,37 @@ export function updateSortAndFilterSettings({
 		}
 	});
 }
+
+/**
+ * Translates a sort or filter option value to its localized label
+ *
+ * This function looks up the corresponding sorting or filtering option by its value
+ * and returns the translated label. Spaces in label keys are converted to underscores
+ * for translation key lookup (e.g., "last modified" becomes "sorting_dropdown.last_modified").
+ *
+ * @param value - The value of the sort/filter option (e.g., 'date', 'changeDate', 'read')
+ * @param t - The i18next translation function
+ * @returns The translated label for the option, or the original value if no match is found
+ *
+ * @example
+ * // For sort option with value 'changeDate'
+ * getTranslatedSortFilterLabel('changeDate', t)
+ * // Returns: t('sorting_dropdown.last_modified', 'last modified')
+ */
+export const getTranslatedSortFilterLabel = (
+	value: string | null | undefined,
+	t: TFunction<'translation', undefined, 'translation'>
+): string => {
+	if (!value) return '';
+	const sortOpt = Object.values(SORTING_OPTIONS).find((opt) => opt.value === value);
+	if (sortOpt) {
+		const translationKey = sortOpt.label.replace(/ /g, '_');
+		return t(`sorting_dropdown.${translationKey}`, sortOpt.label);
+	}
+	const filterOpt = Object.values(FILTER_OPTIONS).find((opt) => opt.value === value);
+	if (filterOpt) {
+		const translationKey = filterOpt.label.replace(/ /g, '_');
+		return t(`sorting_dropdown.${translationKey}`, filterOpt.label);
+	}
+	return value;
+};
