+++ src/views/app/folder-panel/parts/sort-and-filter-header-component.tsx
@@ -21,44 +21,11 @@ import { useTranslation } from 'react-i18next';
 
 import { SORTING_DIRECTION, SORTING_OPTIONS, FILTER_OPTIONS } from '../../../../constants';
 import {
+	getTranslatedSortFilterLabel,
 	parseMessageSortingOptions,
 	updateSortAndFilterSettings
 } from '../../../../helpers/sorting';
 import type { SortOption, FilterOption, SortAndFilterState } from '../../../../types';
-
-/**
- * Translates a sort or filter option value to its localized label
- *
- * This function looks up the corresponding sorting or filtering option by its value
- * and returns the translated label. Spaces in label keys are converted to underscores
- * for translation key lookup (e.g., "last modified" becomes "sorting_dropdown.last_modified").
- *
- * @param value - The value of the sort/filter option (e.g., 'date', 'changeDate', 'read')
- * @param t - The i18next translation function
- * @returns The translated label for the option, or the original value if no match is found
- *
- * @example
- * // For sort option with value 'changeDate'
- * getTranslatedSortFilterLabel('changeDate', t)
- * // Returns: t('sorting_dropdown.last_modified', 'last modified')
- */
-const getTranslatedSortFilterLabel = (
-	value: string | null | undefined,
-	t: TFunction<'translation', undefined, 'translation'>
-): string => {
-	if (!value) return '';
-	const sortOpt = Object.values(SORTING_OPTIONS).find((opt) => opt.value === value);
-	if (sortOpt) {
-		const translationKey = sortOpt.label.replace(/ /g, '_');
-		return t(`sorting_dropdown.${translationKey}`, sortOpt.label);
-	}
-	const filterOpt = Object.values(FILTER_OPTIONS).find((opt) => opt.value === value);
-	if (filterOpt) {
-		const translationKey = filterOpt.label.replace(/ /g, '_');
-		return t(`sorting_dropdown.${translationKey}`, filterOpt.label);
-	}
-	return value;
-};
 
 const isValid = (
 	val: string | undefined,
